+++ GetThePicture.Tests/Copybook/Warpper/CbWarpperTest.cs
@@ -6,7 +6,7 @@ using GetThePicture.Picture.Clause.Utils;
 
 namespace GetThePicture.Tests.Copybook.Wrapper;
 
-public class T30_t(byte[] raw) : CbWrapper(raw)
+public class T30_t(byte[]? raw = null) : CbWrapper(raw)
 {
     // ----------------------------
     // Copybook Address Map
@@ -46,40 +46,40 @@ public class T30_t(byte[] raw) : CbWrapper(raw)
 
     public string StockNo
     {
-        get => (string)this["STOCK-NO"]!;
-        set => this["STOCK-NO"] = value;
+        get => this["STOCK-NO"].Get<string>();
+        set => this["STOCK-NO"].Set(value);
     }
 
     public decimal BullPrice
     {
-        get => (decimal)this["BULL-PRICE"]!;
-        set => this["BULL-PRICE"] = value;
+        get => this["BULL-PRICE"].Get<decimal>();
+        set => this["BULL-PRICE"].Set(value);
     }
 
     public decimal LdcPrice
     {
-        get => (decimal)this["LDC-PRICE"]!;
-        set => this["LDC-PRICE"] = value;
+        get => this["LDC-PRICE"].Get<decimal>();
+        set => this["LDC-PRICE"].Set(value);
     }
 
     public decimal BearPrice
     {
-        get => (decimal)this["BEAR-PRICE"]!;
-        set => this["BEAR-PRICE"] = value;
+        get => this["BEAR-PRICE"].Get<decimal>();
+        set => this["BEAR-PRICE"].Set(value);
     }
 
     public DateOnly LastMthDate
     {
-        get => (DateOnly)this["LAST-MTH-DATE"]!;
-        set => this["LAST-MTH-DATE"] = value;
+        get => this["LAST-MTH-DATE"].Get<DateOnly>();
+        set => this["LAST-MTH-DATE"].Set(value);
     }
 
     // ...
 
     public string StockName
     {
-        get => (string)this["STOCK-NAME"]!;
-        set => this["STOCK-NAME"] = value;
+        get => this["STOCK-NAME"].Get<string>();
+        set => this["STOCK-NAME"].Set(value);
     }
 
     // ...
@@ -106,32 +106,69 @@ public class CbWrapperTest
         Assert.AreEqual(96.95m, T30.LdcPrice);
         Assert.AreEqual(87.3m, T30.BearPrice);
         Assert.AreEqual(new DateOnly(2025, 12, 19), T30.LastMthDate);
-        Assert.AreEqual("0", T30["SETTYPE"]);
-        Assert.AreEqual("0", T30["MARK-W"]);
-        Assert.AreEqual("0", T30["MARK-P"]);
-        Assert.AreEqual("0", T30["MARK-L"]);
-        Assert.AreEqual("00", T30["IND-CODE"]);
-        Assert.AreEqual("", T30["IND-SUB-CODE"]);
-        Assert.AreEqual("0", T30["MARK-M"]);
+        Assert.AreEqual("0", T30["SETTYPE"].Get<string>());
+        Assert.AreEqual("0", T30["MARK-W"].Get<string>());
+        Assert.AreEqual("0", T30["MARK-P"].Get<string>());
+        Assert.AreEqual("0", T30["MARK-L"].Get<string>());
+        Assert.AreEqual("00", T30["IND-CODE"].Get<string>());
+        Assert.AreEqual("", T30["IND-SUB-CODE"].Get<string>());
+        Assert.AreEqual("0", T30["MARK-M"].Get<string>());
         Assert.AreEqual("台泥一永", T30.StockName);
-        Assert.AreEqual((UInt16) 0, T30["MATCH-INTERVAL"]);
-        Assert.AreEqual((UInt32) 0, T30["ORDER-LIMIT"]);
-        Assert.AreEqual((UInt32) 0, T30["ORDERS-LIMIT"]);
-        Assert.AreEqual((UInt16) 0, T30["PREPAY-RATE"]);
-        Assert.AreEqual("0", T30["MARK-S"]);
-        Assert.AreEqual("0", T30["STK-MARK"]);
-        Assert.AreEqual("0", T30["MARK-F"]);
-        Assert.AreEqual("", T30["MARK-DAY-TRADE"]);
-        Assert.AreEqual("0", T30["STK-CTGCD"]);
-        Assert.AreEqual("", T30["FILLER"]);
+        Assert.AreEqual((UInt16) 0, T30["MATCH-INTERVAL"].Get<ushort>());
+        Assert.AreEqual((UInt32) 0, T30["ORDER-LIMIT"].Get<uint>());
+        Assert.AreEqual((UInt32) 0, T30["ORDERS-LIMIT"].Get<uint>());
+        Assert.AreEqual((UInt16) 0, T30["PREPAY-RATE"].Get<ushort>());
+        Assert.AreEqual("0", T30["MARK-S"].Get<string>());
+        Assert.AreEqual("0", T30["STK-MARK"].Get<string>());
+        Assert.AreEqual("0", T30["MARK-F"].Get<string>());
+        Assert.AreEqual("", T30["MARK-DAY-TRADE"].Get<string>());
+        Assert.AreEqual("0", T30["STK-CTGCD"].Get<string>());
+        Assert.AreEqual("", T30["FILLER"].Get<string>());
 
         T30.StockNo = "2330";
         T30.LastMthDate = new DateOnly(2025, 11, 14);
         T30.StockName = "台積電";
-        T30["MARK-DAY-TRADE"] = "X";
+        T30["MARK-DAY-TRADE"].Set("X");
 
         var str = cp950.GetString(T30.Raw);
 
-         Assert.AreEqual(after, str);
+        Assert.AreEqual(after, str);
+    }
+
+    [TestMethod]
+    public void Wrapper_Default_Value()
+    {
+        var T30 = new T30_t();
+
+        var str = cp950.GetString(T30.Raw);
+
+        const string expected = "      00000000000000000000000000000000000                         000000000000000000                ";
+        Assert.AreEqual(expected, str);
+    }
+
+    [TestMethod]
+    [ExpectedException(typeof(KeyNotFoundException))]
+    public void Wrapper_Get_Field_Throw_KeyNotFoundException()
+    {
+        const string s = "11011 00106600000096950000087300020251219000000  0台泥一永        000000000000000000000 0           ";
+        
+        byte[] raw = cp950.GetBytes(s);
+        
+        var T30 = new T30_t(raw);
+
+        T30["SHOW-ME-THE-MONEY"].Get<uint>();
+    }
+
+    [TestMethod]
+    [ExpectedException(typeof(KeyNotFoundException))]
+    public void Wrapper_Set_Field_Throw_KeyNotFoundException()
+    {
+        const string s = "11011 00106600000096950000087300020251219000000  0台泥一永        000000000000000000000 0           ";
+        
+        byte[] raw = cp950.GetBytes(s);
+        
+        var T30 = new T30_t(raw);
+
+        T30["SHOW-ME-THE-MONEY"].Set(100000000m);
     }
 }